requirements.txt
streamlit

streamlit_timeline
pillow